backend/requirements.txt
fastapi
uvicorn
sqlmodel
google-generativeai
psycopg2-binary
arrow